fix: pass all Phase 1 unit tests — reorderPages cast bug + DB test env (#179)
* Initial plan

* fix: reorderPages integer cast + set default DATABASE_URL in vitest config

Agent-Logs-Url: https://github.com/ZebLawrence/flight-app/sessions/683f6d42-1873-49e9-867b-ce1b9c84dc00

---------

+++ src/lib/db/queries/pages.ts
@@ -120,7 +120,7 @@ export async function reorderPages(tenantId: string, orderedIds: string[]): Prom
   for (let i = 0; i < orderedIds.length; i++) {
     caseExpr = sql`${caseExpr} WHEN ${orderedIds[i]}::uuid THEN ${i}`;
   }
-  caseExpr = sql`${caseExpr} END`;
+  caseExpr = sql`(${caseExpr} END)::integer`;
 
   await db
     .update(pages)
